+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes to this project will be documented in this file.
 * Add volunteer Controller and API - NC
 * Add associative table UserBooking - JC
 * Add more users - JC
+* Add Booking API - JC
+* Modify Get Booking API - JC
 
 #### Changed
 * Fix volunteer model - JC